+++ messages/de/voting.php
@@ -12,6 +12,7 @@ return [
     'status_accepted' => 'Angenommen',
     'status_rejected' => 'Abgelehnt',
     'status_quorum_missed' => 'Quorum verfehlt',
+    'status_quorum_reached' => 'Quorum erreicht',
 
     'activity_title' => 'Protokoll',
     'activity_show_all' => 'Komplettes Protokoll anzeigen',